+++ src/admin/admin.service.ts
@@ -20,6 +20,10 @@ export class AdminService {
   constructor(private readonly dataSource: DataSource,
               private readonly getToken: GetToken) {}
   private readonly logger = new Logger(AdminService.name);
+   /**
+   * 접근대상자가 admin인지 권한확인.
+   * @return boolean
+   */
   async checkIsAdmin(headers: Headers): Promise<boolean> {
     try {
       const verified = await this.getToken.getToken(headers);
@@ -42,7 +46,10 @@ export class AdminService {
       );
     }
   }
-
+   /**
+   * 관리자가 확인하지 않은 모든 ebook게시글 list 가져오기
+   * @return page
+   */
   async getUncheckedEbookList(headers: Headers,
                               pageRequest?: PageRequest): Promise<Object> {
     try {
@@ -87,7 +94,12 @@ export class AdminService {
       );
     }
   }
-  async approveEbookList(headers, ebookList): Promise<Object> {
+  /**
+   * 승인 리스트를 받아서 상태 업데이트
+   * 승인인 경우 admincheck=true로
+   *  @return {success:true}
+   */
+  async accepctEbookList(headers, ebookList): Promise<Object> {
     try {
 
       const isAdmin = await this.checkIsAdmin(headers);
@@ -99,6 +111,38 @@ export class AdminService {
                 .set({adminCheck : true})
                 .where("ebook.id IN (:...ids)", {ids : ebookList.accepctedList})
                 .execute();
+        return {success : true};
+      } else {
+        throw new HttpException(
+            {
+              status : HttpStatus.FORBIDDEN,
+              error : 'checkedList에 일반유저접근',
+              success : false,
+            },
+            HttpStatus.FORBIDDEN,
+        );
+      }
+    } catch (err) {
+      throw new HttpException(
+          {
+            status : HttpStatus.INTERNAL_SERVER_ERROR,
+            error : 'checkedList에서 오류발생',
+            success : false,
+          },
+          HttpStatus.INTERNAL_SERVER_ERROR,
+      )
+    }
+  }
+    /**
+   * 비승인 리스트를 받아서 상태 업데이트
+   * 비승인인경우 ban된것으로 간주하였습니다.
+   *   @return {success:true}
+   */
+  async rejectEbookList(headers, ebookList): Promise<Object> {
+    try {
+
+      const isAdmin = await this.checkIsAdmin(headers);
+      if (isAdmin) {
         const rejected =
             await this.dataSource.getRepository(EbookEntity)
                 .createQueryBuilder('ebook')
@@ -121,7 +165,7 @@ export class AdminService {
       throw new HttpException(
           {
             status : HttpStatus.INTERNAL_SERVER_ERROR,
-            error : 'checkedList에서 오류발생',
+            error : 'rejectList에서 오류발생',
             success : false,
           },
           HttpStatus.INTERNAL_SERVER_ERROR,
