+++ src/qna/qna.controller.ts
@@ -25,11 +25,11 @@ import {
   UpdateQnaCommentDto
 } from './dto/qnacomment.dto';
 import {QnaService} from './qna.service';
-
+import { QnaCommentService } from './qnaComment.service';
 @Controller('qna')
 @ApiTags('Qna API')
 export class QnaController {
-  constructor(private QnaService: QnaService) {}
+  constructor(private qnaService: QnaService,private qnaCommentService:QnaCommentService) {}
   private readonly logger = new Logger(QnaController.name);
   @Public()
   @ApiOperation({summary : 'Qna 전체 조회'})
@@ -39,7 +39,7 @@ export class QnaController {
   @Get()
   async getAll(@Query() page: PageRequest, @Headers() headers) {
     this.logger.log('-----GET /Qna');
-    return await this.QnaService.getAll(page);
+    return await this.qnaService.getAll(page);
   }
 
   @ApiOperation({summary : 'Qna 열람'})
@@ -52,9 +52,9 @@ export class QnaController {
   ) {
     this.logger.log('-----GET /Qna/:id');
     if (page) {
-      return await this.QnaService.getOne(id, headers, page);
+      return await this.qnaService.getOne(id, headers, page);
     } else {
-      return await this.QnaService.getOne(id, headers);
+      return await this.qnaService.getOne(id, headers);
     }
   }
   @ApiOperation({summary : 'Admin의 Qna 열람'})
@@ -62,7 +62,7 @@ export class QnaController {
   @Post('/one/:id')
   async getOnebyAdmin(@Param('id') id: number, @Headers() headers) {
     this.logger.log('-----Post /Qna/:id');
-    return await this.QnaService.getOnebyAdmin(id, headers);
+    return await this.qnaService.getOnebyAdmin(id, headers);
   }
 
   @ApiOperation({summary : 'Qna 작성'})
@@ -70,14 +70,14 @@ export class QnaController {
   @Post('/create')
   async create(@Body() createQnaDto: CreateQnaDto, @Headers() headers) {
     this.logger.log('-----POST /Qna');
-    return await this.QnaService.create(createQnaDto, headers);
+    return await this.qnaService.create(createQnaDto, headers);
   }
   @ApiOperation({summary : 'Qna 수정 전 가져오기'})
   @HttpCode(HttpStatus.OK)
   @Get('/getupdate/:id')
   async getUpdate(@Param('id') id: number, @Headers() headers) {
     this.logger.log('-----GET /Qna/getupdate');
-    return await this.QnaService.getUpdate(id, headers);
+    return await this.qnaService.getUpdate(id, headers);
   }
 
   @ApiOperation({summary : 'Qna 수정'})
@@ -85,7 +85,7 @@ export class QnaController {
   @Patch()
   async update(@Body() updateQnaDto: UpdateQnaDto, @Headers() headers) {
     this.logger.log('-----PATCH /Qna');
-    return await this.QnaService.update(updateQnaDto, headers);
+    return await this.qnaService.update(updateQnaDto, headers);
   }
 
   @ApiOperation({summary : 'Qna 삭제'})
@@ -93,7 +93,7 @@ export class QnaController {
   @Delete()
   async delete(@Body() deleteQnaDto: DeleteQnaDto, @Headers() headers) {
     this.logger.log('-----DELETE /Qna');
-    return await this.QnaService.delete(deleteQnaDto, headers);
+    return await this.qnaService.delete(deleteQnaDto, headers);
   }
   @ApiOperation({summary : 'Qna comment 작성'})
   @HttpCode(HttpStatus.CREATED)
@@ -102,7 +102,7 @@ export class QnaController {
                       @Body() createQnaCommentDto: CreateQnaCommentDto,
                       @Headers() headers) {
     this.logger.log('-----Post /Qna/comment/create');
-    return await this.QnaService.createComment(id, createQnaCommentDto,
+    return await this.qnaCommentService.createComment(id, createQnaCommentDto,
                                                headers);
   }
   @ApiOperation({summary : 'Qna  Comment 수정 전 가져오기'})
@@ -111,7 +111,7 @@ export class QnaController {
   async getCommentUpdate(@Param('commentId') commentId: number,
                          @Headers() headers) {
     this.logger.log('-----GET /Qna/getCommentupdate');
-    return await this.QnaService.getCommentUpdate(commentId, headers);
+    return await this.qnaCommentService.getCommentUpdate(commentId, headers);
   }
   @ApiOperation({summary : 'Qna Comment 수정'})
   @HttpCode(HttpStatus.OK)
@@ -119,14 +119,14 @@ export class QnaController {
   async updateComment(@Body() updateQnaCommentDto: UpdateQnaCommentDto,
                       @Headers() headers) {
     this.logger.log('-----PATCH /Qna/Comment');
-    return await this.QnaService.updateComment(updateQnaCommentDto, headers);
+    return await this.qnaCommentService.updateComment(updateQnaCommentDto, headers);
   }
   @ApiOperation({summary : 'Qna Comment 열람'})
   @HttpCode(HttpStatus.OK)
   @Get('/comment/:id')
   async getOneComment(@Param('id') id: number, @Headers() headers) {
     this.logger.log('-----GET /Qna/Comment/:id')
-    return await this.QnaService.getOneComment(id, headers);
+    return await this.qnaCommentService.getOneComment(id, headers);
   }
   @ApiOperation({summary : 'Qna Comment 삭제'})
   @HttpCode(HttpStatus.NO_CONTENT)
@@ -134,6 +134,6 @@ export class QnaController {
   async deleteComment(@Body() deleteQnaCommentDto: DeleteQnaCommentDto,
                       @Headers() headers) {
     this.logger.log('-----DELETE /Qna/Comment ');
-    return await this.QnaService.deleteComment(deleteQnaCommentDto, headers);
+    return await this.qnaCommentService.deleteComment(deleteQnaCommentDto, headers);
   }
 }
