+++ src/qna/qna.service.ts
@@ -19,7 +19,7 @@ import {GetToken} from 'src/utils/GetToken';
 import {Page} from 'src/utils/page';
 import {PageRequest} from 'src/utils/PageRequest';
 import {QueryBuilder, Repository} from 'typeorm';
-
+import { QnaCommentService } from './qnaComment.service';
 import {UpdateQnaDto} from './dto/qna.dto';
 
 @Injectable()
@@ -33,6 +33,7 @@ export class QnaService {
       private readonly jwtService: JwtService,
       private readonly getToken: GetToken,
       private readonly getSearchSql: GetSearchSql,
+      private readonly qnaCommentService: QnaCommentService
   ) {}
 
   async countAll(boardId?: number): Promise<number> {
@@ -251,7 +252,7 @@ export class QnaService {
       const verified = await this.getToken.getToken(headers);
       const check = await this.checkUserandIsSecret(id, verified.userId);
       let page = await this.getQnaPage(id);
-      let comments = await this.getAllComment(id, pageRequest);
+      let comments = await this.qnaCommentService.getAllComment(id, pageRequest);
       if (comments['success']) {
         comments = comments['Page'];
       } else {
@@ -451,211 +452,5 @@ export class QnaService {
                               HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
-  async createComment(id, createQnaCommentDto, headers): Promise<Object> {
-    const verified = await this.getToken.getToken(headers);
-    const qnaCommentData = new QnaCommentEntity();
-    qnaCommentData.title = createQnaCommentDto.title;
-    qnaCommentData.contents = createQnaCommentDto.contents;
-    qnaCommentData.user = verified.userId;
-    qnaCommentData.username = verified.username;
-    qnaCommentData.dateTime = new Date();
-    qnaCommentData.isDeleted = false;
-    qnaCommentData.isModified = false;
-    qnaCommentData.issecret = createQnaCommentDto.issecret;
-    qnaCommentData.qna = id;
-    qnaCommentData.parentId = createQnaCommentDto.parnetId;
-    if (!qnaCommentData.issecret) {
-      qnaCommentData.username = verified.username;
-    }
-    await this.qnaCommentRepository.save(qnaCommentData);
-    return {success : true, status : HttpStatus.OK};
-  }
-  catch(err) {
-    this.logger.error(err);
-    throw new HttpException({
-      status : HttpStatus.INTERNAL_SERVER_ERROR,
-      error : 'QnaComment 생성중 오류발생',
-      success : false,
-    },
-                            HttpStatus.INTERNAL_SERVER_ERROR)
-  }
-  async getAllComment(boarId: number,
-                      pageRequest?: PageRequest): Promise<Object> {
-    try {
-      let offset = 10;
-      let limit = 10;
-      let pageSize = 10;
-      if (pageRequest) {
-        offset = pageRequest.getOffset();
-        limit = pageRequest.getLimit();
-        pageSize = pageRequest.pageSize;
-      }
-      const count = await this.countAll(boarId);
-      const page = await this.qnaCommentRepository.query(
-          `select  id, title,username, "dateTime","issecret" from "qnaComment" as q where q."isDeleted"=false and q."ban"=false order by q."parentId" desc offset ${
-              offset} limit ${limit}`);
-      const returnPage = new Page(count, pageSize, page);
-      return {success : true, Page : returnPage};
-    } catch (err) {
-      this.logger.error(err);
-      throw new HttpException({
-        status : HttpStatus.INTERNAL_SERVER_ERROR,
-        error : 'Qna 전체조회중 오류발생',
-        success : false,
-      },
-                              HttpStatus.INTERNAL_SERVER_ERROR)
-    }
-  }
-  async getOneComment(id: number, headers: Headers, page?): Promise<Object> {
-    try {
-      const isComment = true;
-      const verified = await this.getToken.getToken(headers);
-      const check =
-          await this.checkUserandIsSecret(id, verified.userId, isComment);
-      const page = await this.qnaCommentRepository.query(
-          `select  id, title, "dateTime",username,contents from "qnaComment" where "id"=${
-              id} and "ban"=false and "isDeleted"=false`);
-      if (!page) {
-        throw new Error('Qna.getOneComment에서 삭제된 Qna에 접근')
-      }
-      return {
-        success : true,
-        page,
-        check : check['check'],
-        status : HttpStatus.OK
-      };
-
-    } catch (err) {
-      this.logger.error(err);
-      throw new HttpException({
-        status : HttpStatus.INTERNAL_SERVER_ERROR,
-        error : 'Qna 조회중 오류발생',
-        success : false,
-      },
-                              HttpStatus.INTERNAL_SERVER_ERROR)
-    }
-  }
-  async getCommentUpdate(id, headers): Promise<Object> {
-    try {
-      const update = await this.getOneComment(id, headers);
-      const checkIsOwner = await this.checkIsOwner(update['check']);
-      if (!checkIsOwner)
-        throw new Error('Qna.getupdate에 무권한접근');
-
-      return {
-        success : true,
-        status : update['status'],
-        page : update['page'],
-        check : update['check']
-      };
-
-    } catch (err) {
-      this.logger.error(err);
-      throw new HttpException({
-        status : HttpStatus.INTERNAL_SERVER_ERROR,
-        error : 'Qna 업데이트 전 오류발생',
-        success : false,
-      },
-                              HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-  /**
-   *qna 수정하기
-   *
-   *@param updateQnaDto,headers
-   *@return  { success};
-   */
-  async updateComment(updateQnaCommentDto, headers): Promise<Object> {
-    try {
-      this.logger.log(updateQnaCommentDto);
-      this.logger.log(headers);
-      const isComment = true;
-      const verified = await this.getToken.getToken(headers);
-      const check = await this.checkUserandIsSecret(
-          updateQnaCommentDto.qnaCommentId, verified.userId, isComment);
-
-      const isOwner = await this.checkIsOwner(check['check']);
-
-      if (!isOwner)
-        throw new Error('Qna.Commentupdate 에 무권한접근');
-
-      const res = await this.updateQnaComment(updateQnaCommentDto);
-
-      return res;
-
-    } catch (err) {
-      this.logger.error(err);
-      throw new HttpException({
-        status : HttpStatus.INTERNAL_SERVER_ERROR,
-        error : 'Qna Comment업데이트 중 오류발생',
-        success : false,
-      },
-                              HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  /**
-   * @param updateQnaCommentDto
-   * @returns {Success,status}
-   */
-  async updateQnaComment(updateQnaCommentDto) {
-    try {
-      await this.qnaCommentRepository
-          .createQueryBuilder("qnaComment") // 이거 안에 딱히 상관은 없을건데
-                                            // 테이블 명으로 바꿔주세요
-          .update()
-          .set(
-              {
-                title : updateQnaCommentDto.title,
-                contents : updateQnaCommentDto.contents,
-                isModified : true,
-                issecret : updateQnaCommentDto.isSecret,
-              },
-              )
-          .where('id = :id', {
-            id : updateQnaCommentDto.qnaCommentId
-          }) // camel 표기법 꼭 지켜주세요 QnaId입니다
-          .execute();
-      return {success : true, status : HttpStatus.OK};
-    } catch (err) {
-      this.logger.error(err);
-      throw new HttpException({
-        status : HttpStatus.INTERNAL_SERVER_ERROR,
-        error : 'Qna Comment 업데이트 중 오류발생',
-        success : false,
-      },
-                              HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-  async deleteComment(deleteQnaCommentDto, headers): Promise<Object> {
-    try {
-      const isComment = true;
-      const verified = await this.getToken.getToken(headers);
-      const isUser = await this.checkUserandIsSecret(
-          deleteQnaCommentDto.qnaCommentId, verified.userId, isComment);
-      if (!isUser['check']['isOwner'])
-        throw new Error('Qna.delete Comment 에서 무권한접근');
-
-      await this.qnaRepository.createQueryBuilder('qnaComment')
-          .update()
-          .set(
-              {
-                isDeleted : true,
-              },
-              )
-          .where('id = :id', {id : deleteQnaCommentDto.qnaCommentId})
-          .execute();
-
-      return {success : true, status : HttpStatus.NO_CONTENT};
-
-    } catch (err) {
-      this.logger.error(err);
-      throw new HttpException({
-        status : HttpStatus.INTERNAL_SERVER_ERROR,
-        error : 'Qna Comment 삭제 중 오류발생',
-        success : false,
-      },
-                              HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
+  
 }
